backend/requirements.txt
fastapi==0.115.12
SQLAlchemy==2.0.40
python-dotenv==1.1.0
pytz==2025.2
starlette==0.46.1
uvicorn==0.34.0
gunicorn==23.0.0